moved types to one file
