--- src/api/types/params.ts
@@ -1,6 +1,0 @@
-export type ManageRolesParams = {
-  guildId: string;
-  userId: string;
-  roleIds: string[];
-  message: string;
-};
